+++ src/app/api/trigger/route.ts
@@ -2,13 +2,13 @@ import { client } from "@/trigger";
 import { createAppRoute } from "@trigger.dev/nextjs";
 
 import "@/jobs/auth-verification-email";
-import "@/jobs/data-room-share-email";
 import "@/jobs/esign-confirmation-email";
 import "@/jobs/esign-email";
 import "@/jobs/esign-pdf";
 import "@/jobs/member-inivite-email";
 import "@/jobs/password-reset-email";
-import "@/jobs/update-share-email";
+import "@/jobs/share-data-room-email";
+import "@/jobs/share-update-email";
 
 //this route is used to send and receive data with Trigger.dev
 export const { POST, dynamic } = createAppRoute(client);
